requirements.txt
streamlit==1.29.0
pandas==2.1.3
requests==2.31.0
python-dotenv==1.0.0
google-generativeai==0.3.1
matplotlib==3.8.2
pyarrow==14.0.1